requirements.txt
pytest==7.2.0
PyHamcrest==2.1.0
pytest-asyncio==0.23.5
websockets==12.0